Fixes #22; Fixed Pie Transitions
This addresses an issue where pie charts were getting "out of sync" with the data model and not removing / repositioning old arcs. The fix was to simple select all and remove before rendering the arcs in the SVG. Unfortunately this does little to address #1 since this approach doesn't allow for smooth transitions. Seems better to ensure it is working correctly now and worry about smooth transitions later.