Closes #32; Significantly Improved QueryCSS
QueryCSS now takes the full DOM tree into account when collecting styles from reference elements. Previously the query was limited to only the containing element of the canvas and completely missed styles that referenced parents further up the DOM tree. Now we build a full tree, up to (but not including) the body tag and use it query styles.