Closes #62; Removed Dependance on jQuery
Completely removed jQuery as a dependancy of Epoch by creating a mini querying system for the library. Basically I hunted down all the ways that jQuery was being used and ensured that we mirrored the functionality using regular JavaScript.

The new querying system, named `Epoch.Query` is where most of the heavy lifting takes place. In places where mirroring functionality seemed out of scope, or could be easily achieved with a simple call or two, I opted not to mirror functionality. One example is that of generating elements directly from html strings.

This change touches quite a few places in the library and fundamentally changes the way we handle DOM querying and manipulation. As such, it would be prudent to get a few eyes on these changes before merging it into master.