Watch domain and range change in Epoch.Time.Plot
Currently users are not able to change domain and range for a time plot because the variables are not watched. We need to watch option.domain and option.range and call @draw() when they are changed. Interestingly, the two variables are watched in basic.coffee, but not in time.coffee.