Remove jQuery from docs; addresses #139 and #238
This addresses issues #139 and #238, which complained that
the examples in the documentation require jQuery to work.

The jQuery dependence has been removed by pull request #84.
According to issue #91, there should be no jQuery requirement
in the documentation either, and indeed the documentation does
not have `<script src="js/jquery.min.js"></script>` anywhere.

Nevertheless, all the examples in the documentation still
use jQuery. As it is now 2018, nobody uses jQuery anymore,
and #145 shows how to use Epoch without jQuery, this commit
and pull request removes jQuery from all the documentation.